Remove side-effects from declarative modules
