+++ services/server/src/server/controllers/verification/etherscan/stateless/etherscan.stateless.handlers.ts
@@ -4,19 +4,22 @@ import {
   getMetadataFromCompiler,
   processRequestFromEtherscan,
 } from "../etherscan.common";
-import { checkSupportedChainId } from "../../../../../sourcify-chains";
 import { getResponseMatchFromMatch } from "../../../../common";
 import { createCheckedContract } from "../../verification.common";
-import { sourcifyChainsMap } from "../../../../../sourcify-chains";
 import logger from "../../../../../common/logger";
+import { ChainRepository } from "../../../../../sourcify-chain-repository";
+import { ISolidityCompiler } from "@ethereum-sourcify/lib-sourcify";
 
 export async function verifyFromEtherscan(req: Request, res: Response) {
-  checkSupportedChainId(req.body.chain);
+  const chainRepository = req.app.get('chainRepository') as ChainRepository;
+  chainRepository.checkSupportedChainId(req.body.chain);
+
+  const solc = req.app.get('solc') as ISolidityCompiler;
 
   const chain = req.body.chain as string;
   const address = req.body.address;
   const apiKey = req.body.apiKey;
-  const sourcifyChain = sourcifyChainsMap[chain];
+  const sourcifyChain = chainRepository.supportedChainMap[chain];
 
   logger.info("verifyFromEtherscan", { chain, address, apiKey });
 
@@ -24,13 +27,14 @@ export async function verifyFromEtherscan(req: Request, res: Response) {
     await processRequestFromEtherscan(sourcifyChain, address, apiKey);
 
   const metadata = await getMetadataFromCompiler(
+    solc,
     compilerVersion,
     solcJsonInput,
     contractName,
   );
 
   const mappedSources = getMappedSourcesFromJsonInput(solcJsonInput);
-  const checkedContract = createCheckedContract(metadata, mappedSources);
+  const checkedContract = createCheckedContract(solc, metadata, mappedSources);
 
   const match = await req.services.verification.verifyDeployed(
     checkedContract,
