+++ services/server/src/server/controllers/verification/session-state/session-state.handlers.ts
@@ -1,21 +1,18 @@
 import { Response, Request } from "express";
 import {
-  ContractWrapperMap,
   FILE_ENCODING,
-  SendableContract,
   addRemoteFile,
   checkContractsInSession,
   extractFiles,
   getSessionJSON,
-  isVerifiable,
   saveFilesToSession,
   verifyContractsInSession,
 } from "../verification.common";
 import {
+  ISolidityCompiler,
   PathBuffer,
   PathContent,
   getIpfsGateway,
-  isEmpty,
   performFetch,
 } from "@ethereum-sourcify/lib-sourcify";
 import { BadRequestError } from "../../../../common/errors";
@@ -41,12 +38,14 @@ export async function addInputFilesEndpoint(req: Request, res: Response) {
     return { path: pb.path, content: pb.buffer.toString(FILE_ENCODING) };
   });
 
+  const solc = req.app.get("solc") as ISolidityCompiler;
   const dryRun = Boolean(req.query.dryrun);
   const session = req.session;
   const newFilesCount = saveFilesToSession(pathContents, session);
   if (newFilesCount) {
-    await checkContractsInSession(session);
+    await checkContractsInSession(solc, session);
     await verifyContractsInSession(
+      solc,
       session.contractWrappers,
       session,
       req.services.verification,
@@ -78,6 +77,7 @@ export async function restartSessionEndpoint(req: Request, res: Response) {
 export async function addInputContractEndpoint(req: Request, res: Response) {
   const address: string = req.body.address;
   const chainId: string = req.body.chainId;
+  const solc = req.app.get("solc") as ISolidityCompiler;
 
   const sourcifyChain = req.services.verification.supportedChainsMap[chainId];
 
@@ -118,9 +118,10 @@ export async function addInputContractEndpoint(req: Request, res: Response) {
 
   const newFilesCount = saveFilesToSession(pathContents, session);
   if (newFilesCount) {
-    await checkContractsInSession(session);
+    await checkContractsInSession(solc, session);
     // verifyValidated fetches missing files from the contract
     await verifyContractsInSession(
+      solc,
       session.contractWrappers,
       session,
       req.services.verification,
