+++ services/server/src/server/services/compiler/lambda-with-fallback/SolcLambdaWithLocalFallback.ts
@@ -8,8 +8,24 @@ import { SolcLambda, LambdaResponseLimitExceeded } from "../lambda/SolcLambda";
 import { SolcLocal } from "../local/SolcLocal";
 
 export class SolcLambdaWithLocalFallback implements ISolidityCompiler {
-  private solcLambda = new SolcLambda();
-  private solcLocal = new SolcLocal();
+  private solcLambda;
+  private solcLocal;
+
+  constructor(
+    awsRegion: string,
+    awsAccessKeyId: string,
+    awsSecretAccessKey: string,
+    lambdaCompilerFunctionName: string = "compile",
+    repoPath: string,
+  ) {
+    this.solcLambda = new SolcLambda(
+      awsRegion,
+      awsAccessKeyId,
+      awsSecretAccessKey,
+      lambdaCompilerFunctionName,
+    );
+    this.solcLocal = new SolcLocal(repoPath);
+  }
 
   public async compile(
     version: string,
