+++ services/server/test/unit/utils/contract-creation-util.spec.ts
@@ -1,9 +1,11 @@
 import chai from "chai";
 import { getCreatorTx } from "../../../src/server/services/utils/contract-creation-util";
-import { sourcifyChainsArray } from "../../../src/sourcify-chains";
+import { sourcifyChainsMap } from "../../../src/sourcify-chains";
+import { ChainRepository } from "../../../src/sourcify-chain-repository";
 
 describe("contract creation util", function () {
   it("should run getCreatorTx with chainId 40", async function () {
+    const sourcifyChainsArray = new ChainRepository(sourcifyChainsMap).sourcifyChainsArray;
     const sourcifyChain = sourcifyChainsArray.find(
       (sourcifyChain) => sourcifyChain.chainId === 40,
     );
@@ -41,6 +43,7 @@ describe("contract creation util", function () {
   // });
 
   it("should run getCreatorTx with chainId 83", async function () {
+    const sourcifyChainsArray = new ChainRepository(sourcifyChainsMap).sourcifyChainsArray;
     const sourcifyChain = sourcifyChainsArray.find(
       (sourcifyChain) => sourcifyChain.chainId === 83,
     );
@@ -59,6 +62,7 @@ describe("contract creation util", function () {
   });
 
   it("should run getCreatorTx with chainId 335", async function () {
+    const sourcifyChainsArray = new ChainRepository(sourcifyChainsMap).sourcifyChainsArray;
     const sourcifyChain = sourcifyChainsArray.find(
       (sourcifyChain) => sourcifyChain.chainId === 335,
     );
@@ -77,6 +81,7 @@ describe("contract creation util", function () {
   });
 
   it("should run getCreatorTx with regex for new Blockscout", async function () {
+    const sourcifyChainsArray = new ChainRepository(sourcifyChainsMap).sourcifyChainsArray;
     const sourcifyChain = sourcifyChainsArray.find(
       (sourcifyChain) => sourcifyChain.chainId === 100,
     );
@@ -95,6 +100,7 @@ describe("contract creation util", function () {
   });
 
   it("should run getCreatorTx with regex for old Blockscout", async function () {
+    const sourcifyChainsArray = new ChainRepository(sourcifyChainsMap).sourcifyChainsArray;
     const sourcifyChain = sourcifyChainsArray.find(
       (sourcifyChain) => sourcifyChain.chainId === 57,
     );
@@ -118,6 +124,7 @@ describe("contract creation util", function () {
       return;
     }
 
+    const sourcifyChainsArray = new ChainRepository(sourcifyChainsMap).sourcifyChainsArray;
     const sourcifyChain = sourcifyChainsArray.find(
       (sourcifyChain) => sourcifyChain.chainId === 1,
     );
@@ -136,6 +143,7 @@ describe("contract creation util", function () {
   });
 
   it("should run getCreatorTx with nexusApi for Nexus", async function () {
+    const sourcifyChainsArray = new ChainRepository(sourcifyChainsMap).sourcifyChainsArray;
     const sourcifyChain = sourcifyChainsArray.find(
       (sourcifyChain) => sourcifyChain.chainId === 23294,
     );
